Set LUVI_VERSION manually
Do not use git to set LUVI_VERSION as it may be not available on the build
system, e.g. by just downloading the github source tarball.

Furthermore, if luvi is build by Buildroot cloned with git, `git describe` does
return the tag from Buildroot as luvi version which is obviously wrong.