LuaJITAddExecutable: pass LJ_BYTECODE_OPTS as bytecode options to LuaJIT
Always passing the '-a' option as a bytecode option has a severe drawback: we
need to pass always a target architecture to LuaJIT and cannot use the
native build used by LuaJIT if no '-a' is passed.

Fix this by passing the '-a' option conditionally by using the variable
LJ_BYTECODE_OPTS. This also allows as to pass other bytecode options through
the environment.