CMake: tweak setting luvi libraries
When building luvi against a shared luv library we want to tell the linker
where to look for libraries. That's why we'll set the luvi libraries to the
values found by find_package(). This is needed for adding a proper RPATH
value to the luvi executable as done in a follow-up patch.