CMake: add RPATH to executable
By default CMake will link the executables and shared libraries with full RPATH
to all used libraries in the build tree. When installing, it will clear the
RPATH of these targets so they are installed with an empty RPATH.

To ensure that the required libraries are always found independent from
LD_LIBRARY_PATH and the install location we have to include RPATH in the
executable.

With RPATH set we will be able to execute luvi from the build tree and luvi will
find the shared libraries in the build tree and also the shared libraries
outside of the build.

This is very important for cross building luvi applications where you run a luvi
executable on your host machine which is built against host libraries installed
in a sysroot directory. If we are not setting RPATH properly, luvi will use the
shared libraries from your host machine, which is not what we want.

Another (none cross-compilation) scenary is where you, let's say, build a shared
luv and install in into a local directory, e.g. "~/usr", instead of into the
system.

To install luvi let luvi find all necessary libraries we build luvi with:
  $ cmake -DCMAKE_INSTALL_PREFIX="~/usr" -DLUA_PATH=/usr/share/luajit-2.0.4/?.lua -DWithSharedLibluv=ON

This will show you:
  [100%] Linking C executable luvi
  [100%] Built target luvi
  [joerg@nzxt luvi]$ make install
  [100%] Built target luvi
  Install the project...
  -- Install configuration: "Release"
  -- Installing: /home/joerg/usr/bin/luvi
  -- Set runtime path of "/home/joerg/usr/bin/luvi" to "/home/joerg/usr/lib"

Note that CMake automatically changes the runtime path with all necessary
libraries necessary for running luvi.

We can check this with readelf:
  $ readelf -d ~/usr/bin/luvi | grep rpath
   0x000000000000000f (RPATH)              Library rpath: [/home/joerg/usr/lib]