Add support for zlib compression with miniz (#163)
This allows users of luvi to use zlib compression without the zlib
module, meaning that it is technically no longer necessary.

This has been tested on Debian buster (current testing) on a 64-bit AMD
processor, where it appears to work correctly. I expect it to work
correctly on all current targets.