Ensure that using the -v or -h flags always returns with EXIT_SUCCESS
Returning -1 indicates an error, assuming the usual convention for exit codes is followed. Seeing how main.c returns -1 in case of errors, this appears to be the case.

Specifically, using either of the two flags in a script or CI workflow will cause it to mistakenly assume an error has occured and fail as a result.