build: audit openssl configure, update to openssl 3.4.0
Updates to the latest stable openssl, this series
has support past the end of the current LTS, and
we have no reason to stay on 3.0.

Now autodetects an existing openssl installation
and uses it for `--openssldir`.

Removes existing `no-*` options that have been
default for a long time.

Adds `no-pinshared` because we're using a static
openssl and don't need it to pin the executable.

Replaces `no-unit-test` (a default) with `no-tests`
to ensure tests are not compiled.

Removes `no-stdio` because `FILE` may be useful,
and it allows generating the openssl `apps/*` for
debugging.

Adds `no-makedepend` because we're running this
once and it roughly halves the number of cc calls.
Which matters quite a bit on Windows.